LocationRestrictedEmwfOptionsView -> LocationRestrictedEmwfOptions, Move CaseListFilterOptions to api module, extract LocationRestrictedEmwfOptions Mixin [skip ci]
