+++ corehq/apps/reports/filters/case_list.py
@@ -1,11 +1,8 @@
 from django.utils.translation import ugettext as _
 
 from dimagi.utils.decorators.memoized import memoized
-from corehq.apps.locations.models import LocationType, SQLLocation
 
 from .users import ExpandedMobileWorkerFilter, EmwfUtils
-from .api import EmwfOptionsView
-from corehq.apps.locations.permissions import location_safe
 
 
 class CaseListFilterUtils(EmwfUtils):
@@ -68,59 +65,3 @@ class CaseListFilter(ExpandedMobileWorkerFilter):
                 self.request.domain
             )
             return map(self.utils.location_tuple, all_locations)
-
-
-@location_safe
-class CaseListFilterOptions(EmwfOptionsView):
-
-    @property
-    @memoized
-    def utils(self):
-        return CaseListFilterUtils(self.domain)
-
-    def get_users(self, query, start, size):
-        users = (self.user_es_query(query)
-                 .fields(['_id', 'username', 'first_name', 'last_name', 'doc_type'])
-                 .start(start)
-                 .size(size)
-                 .sort("username.exact"))
-        if not self.request.can_access_all_locations:
-            user_location_id = self.request.couch_user.get_location_id(self.domain)
-            all_location_ids = SQLLocation.objects.get_locations_and_children_ids([user_location_id])
-            users = users.location(all_location_ids)
-
-        return [self.utils.user_tuple(u) for u in users.run().hits]
-
-    @property
-    def data_sources(self):
-        print 'fetching data source for filters in view'
-        locations_own_cases = (LocationType.objects
-                               .filter(domain=self.domain, shares_cases=True)
-                               .exists())
-        sources = []
-        if locations_own_cases:
-            sources.append((self.get_locations_size, self.get_locations))
-        if self.request.can_access_all_locations:
-            sources.append((self.get_static_options_size, self.get_static_options))
-            sources.append((self.get_sharing_groups_size, self.get_sharing_groups))
-            sources.append((self.get_groups_size, self.get_groups))
-        # appending this in the end to avoid long list of users delaying
-        # locations, groups etc in the list on pagination
-        sources.append((self.get_users_size, self.get_users))
-        return sources
-
-    def get_sharing_groups_size(self, query):
-        return self.group_es_query(query, group_type="case_sharing").count()
-
-    def get_sharing_groups(self, query, start, size):
-        groups = (self.group_es_query(query, group_type="case_sharing")
-                  .fields(['_id', 'name'])
-                  .start(start)
-                  .size(size)
-                  .sort("name.exact"))
-        return map(self.utils.sharing_group_tuple, groups.run().hits)
-
-    def get_locations_query(self, query):
-        return (SQLLocation.active_objects
-                .filter_path_by_user_input(self.domain, query)
-                .accessible_to_user(self.request.domain, self.request.couch_user))
