+++ corehq/apps/reports/filters/urls.py
@@ -1,12 +1,11 @@
 from django.conf.urls import url
 
-from .api import EmwfOptionsView, LocationRestrictedEmwfOptionsView
-from .case_list import CaseListFilterOptions
+from .api import EmwfOptionsView, LocationRestrictedEmwfOptions, CaseListFilterOptions
 from .location import LocationGroupFilterOptions
 
 urlpatterns = [
    url(r'^emwf_options/$', EmwfOptionsView.as_view(), name='emwf_options'),
-   url(r'^new_emwf_options/$', LocationRestrictedEmwfOptionsView.as_view(), name='new_emwf_options'),
+   url(r'^new_emwf_options/$', LocationRestrictedEmwfOptions.as_view(), name='new_emwf_options'),
    url(r'^case_list_options/$', CaseListFilterOptions.as_view(),
        name='case_list_options'
        ),
